Extract components that uikit dynamicly reload and create the non ssr version
