+++ components/utils/Nav.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const Nav = ({ navbar, items }) => (
+  <ul className={navbar ? 'uk-navbar-nav' : 'uk-nav uk-nav-default'}>
+    {items.map((value, index) => (
+      <li key={index}>{value}</li>
+    ))}
+  </ul>
+);
+Nav.propTypes = {
+  navbar: PropTypes.bool,
+  items: PropTypes.arrayOf(PropTypes.element).isRequired,
+};
+Nav.defaultProps = {
+  navbar: false,
+};
+
+export default Nav;
