+++ components/utils/index.js
@@ -1,5 +1,3 @@
 export * from './card';
 export * from './links';
-export * from './navs';
 export * from './layout';
-export * from './elements';
